+++ Frontend/pet-app/src/app/page.tsx
@@ -1,3 +1,9 @@
+import { NuevoGasto } from "@/components/NuevoGasto/NuevoGasto";
+
 export default function Home() {
-  return <>ACA TENDRIA QUE IR EL LOGIN O REGISTRO</>;
+  return (
+    <>
+      <NuevoGasto />
+    </>
+  );
 }
