+++ Frontend/pet-app/src/components/containerMain/containerMain.module.css
@@ -4,6 +4,6 @@
   left: 270px;
   top: 60px;
   width: calc(100vw - 270px);
-  height: calc(100vh - 60px);
   background-color: red;
+  height: fit-content;
 }
