+++ Frontend/pet-app/src/components/navbarSuperior/navbarSup.module.css
@@ -7,10 +7,11 @@
   display: flex;
   justify-content: flex-end;
   z-index: 10;
+  color: white;
 }
 
 .containerBtns {
   display: flex;
   gap: 15px;
-  margin-right: 20px;
+  margin-right: 60px;
 }
