+++ Frontend/pet-app/src/components/navbarSuperior/navbarSup.tsx
@@ -4,7 +4,7 @@ export const NavbarSup = () => {
   return (
     <div className={styles.containerNavSup}>
       <div className={styles.containerBtns}>
-        <button>NEW</button>
+        <button>NOTI</button>
         <button>USER</button>
       </div>
     </div>
